updates login and logout actions
Why is this necessary?

Logging in did not redirect to user dashboard
and logging out was not updating dom when user-token
was deleted.

uses a stored state to check if user logged in
and that is much more effective.

also redefined a promise on login to check for success
or rejection